simplify things a bit: no esmodules (because they're always deferred) jam all client js into a single file (might reverse this eventually) no static server removed extra config
