+++ lib/client-scripts.js
@@ -3,43 +3,33 @@
 var fs = require("fs");
 var path = require("path");
 var Transform = require("stream").Transform;
-var serveStatic = require("serve-static");
-var finalhandler = require("finalhandler");
 var contentTypes = require("./content-types");
 
 module.exports = function (config) {
   const clientDir = "client";
-  const clientDirAbsolute = path.join(__dirname, clientDir);
-  const staticPrefix = config.prefix + clientDir;
+  const clientScriptPath = config.prefix + clientDir + "/unblocker.js";
 
-  // todo: enable immtable caching if NODE_ENV is production
-  const handleStaticRequest = serveStatic(clientDirAbsolute, {
-    index: false, // don't serve up index.html files for /
-    fallthrough: false, // anything missing is a 404
-  });
-
-  function server(data) {
-    const req = data.clientRequest;
-    const res = data.clientResponse;
-    const url = req.url;
-    if (url.startsWith(staticPrefix) && url.length > staticPrefix.length) {
-      // trim the URL down to be relative to the client dir
-      req.originalUrl = url;
-      req.url = url.substr(staticPrefix.length);
-      // todo: allow handlers and middleware to be async, then drop finalhandler
-      handleStaticRequest(req, res, finalhandler(req, res));
-      return true; // true = this request has been handled, no need to process it further
-    }
+  // todo: async
+  function generateClientScript() {
+    return fs
+      .readFileSync(path.join(__dirname, clientDir, "unblocker.template.js"))
+      .toString()
+      .replace("PREFIX", config.prefix);
   }
 
-  const CLIENT_SCRIPT_INIT =
-    '<script type="module">\n' +
-    fs
-      .readFileSync(path.join(__dirname, clientDir, "init.template.mjs"))
-      .toString()
-      .replaceAll("CLIENT_PATH", staticPrefix)
-      .replace("PREFIX", config.prefix) +
-    "</script>";
+  function server(req, res, next) {
+    if (req.url === clientScriptPath) {
+      // todo: make this better, at least for NODE_ENV=production
+      res.writeHead(200, {
+        "content-type": "application/javascript; charset=UTF-8",
+      });
+      res.end(generateClientScript());
+      return;
+    }
+    next();
+  }
+
+  const CLIENT_SCRIPT_INIT = `<script src="${clientScriptPath}"></script>`;
 
   function createStream() {
     return new Transform({
