+++ lib/unblocker.js
@@ -30,7 +30,6 @@ Unblocker.websockets = require("./websockets.js");
 var defaultConfig = {
   prefix: "/proxy/",
   host: null, // can be used to override the url used in redirects
-  localHandlers: [],
   requestMiddleware: [],
   responseMiddleware: [],
   standardMiddleware: true,
@@ -40,7 +39,7 @@ var defaultConfig = {
   ),
 };
 
-function expandConfig(config) {
+function Unblocker(config) {
   _.defaults(config, defaultConfig);
 
   if (config.prefix.substr(-1) != "/") {
@@ -81,10 +80,10 @@ function expandConfig(config) {
       metaRobots,
     ].concat(config.responseMiddleware, [contentLength]);
 
+    var clientScripts;
     if (config.clientScripts) {
       // insert clientScripts after the urlPrefixer
-      var clientScripts = Unblocker.clientScripts(config);
-      config.localHandlers.unshift(clientScripts.server);
+      clientScripts = Unblocker.clientScripts(config);
       const position = config.responseMiddleware.indexOf(urlPrefixer) + 1;
       config.responseMiddleware.splice(position, 0, clientScripts.injector);
     }
@@ -101,11 +100,6 @@ function expandConfig(config) {
       "response"
     );
   }
-  return config;
-}
-
-function Unblocker(config) {
-  config = expandConfig(config);
 
   debug("Unblocker initialized, config: ", config);
 
@@ -247,47 +241,46 @@ function Unblocker(config) {
     return data;
   }
 
-  function handleLocally(data) {
-    return config.localHandlers.some(
-      (handler) => handler(data) || data.clientResponse.headersSent
-    );
-  }
+  // todo: make this less confusing
+  const clientScriptsServer = config.clientScripts
+    ? clientScripts.server
+    : (req, res, next) => next();
 
   // regular web requests
   function handleRequest(clientRequest, clientResponse, next) {
     const data = initData(clientRequest, clientResponse);
 
-    if (handleLocally(data)) {
-      return;
-    }
+    clientScriptsServer(clientRequest, clientResponse, (err) => {
+      if (err) return next(err);
 
-    clientResponse.redirectTo = redirectTo.bind(
-      redirectTo,
-      clientRequest,
-      clientResponse
-    );
+      clientResponse.redirectTo = redirectTo.bind(
+        redirectTo,
+        clientRequest,
+        clientResponse
+      );
 
-    if (!next) {
-      next = function () {
-        clientResponse.writeHead(400);
-        clientResponse.end("Unable to process request");
-      };
-    }
-
-    const formatted = data.url;
-    const raw = data.rawUrl;
-
-    if (formatted) {
-      // If the raw URL isn't quite right, but we can figure it out, redirect to the correct URL.
-      // Special exception for cases where routers collapsed slashes (see #130)
-      if (formatted !== raw && formatted.replace("://", ":/") !== raw) {
-        return clientResponse.redirectTo(formatted);
+      if (!next) {
+        next = function () {
+          clientResponse.writeHead(400);
+          clientResponse.end("Unable to process request");
+        };
       }
 
-      proxy(data, next);
-    } else {
-      next();
-    }
+      const formatted = data.url;
+      const raw = data.rawUrl;
+
+      if (formatted) {
+        // If the raw URL isn't quite right, but we can figure it out, redirect to the correct URL.
+        // Special exception for cases where routers collapsed slashes (see #130)
+        if (formatted !== raw && formatted.replace("://", ":/") !== raw) {
+          return clientResponse.redirectTo(formatted);
+        }
+
+        proxy(data, next);
+      } else {
+        next();
+      }
+    });
   }
 
   // websocket support
@@ -319,4 +312,3 @@ function Unblocker(config) {
 
 module.exports = Unblocker;
 module.exports.defaultConfig = defaultConfig;
-module.exports.expandConfig = expandConfig;
